changed to double and added infantry bonus from artillery

+++ Assets/Scripts/Odds.cs
@@ -1,16 +1,16 @@
 public class Odds {
 	
-	public float Success { get; private set; }
-	public float Failure { get; private set; }
+	public double Success { get; private set; }
+	public double Failure { get; private set; }
 	public bool Active { get; set; }
 
 	public Odds(int roll) {
-		this.Success = roll / 6.0f;
-		this.Failure = (6 - roll) / 6.0f;
+		this.Success = roll / 6.0;
+		this.Failure = (6 - roll) / 6.0;
 		this.Active = false;
 	}
 
-	public float GetOdds() { 
+	public double GetOdds() { 
 		return this.Active ? this.Success : this.Failure;
 	}
